rust/python/requirements.txt
blinker==1.6.2
cachetools==5.3.1
click==8.1.6
Flask==2.3.2
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
isodate==0.6.1
itsdangerous==2.1.2
joblib==1.3.1
nltk==3.8.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dotenv==1.0.0
regex==2023.6.3
rsa==4.9
tqdm==4.65.0
uritemplate==4.1.1
Werkzeug==2.3.6
youtube-transcript-api==0.6.1
